solve bug, hapus timkhus tapi absen tidak terhapus

+++ app/Http/Livewire/Desktop/ManualTimkhu.php
@@ -160,7 +160,14 @@ class ManualTimkhu extends Component
     public function delete($id)
     {
         $toDelete=Timkhu::find($id);
-        $toDelete->delete();
+        dd($toDelete->kegiatan);
+
+        // cek jika masih punya absen tidak bisa hapus
+        if($toDelete->absensi->count() > 0)
+            return $this->emit('swalMessageError', 'Tidak bisa menghapus tim yang masih memiliki absensi');
+        else
+            $toDelete->delete();
+            
         // $this->mount();
     }
